feat:add service

+++ src/main/java/com/github/acnaweb/study_apir/model/Product.java
@@ -1,18 +1,18 @@
 package com.github.acnaweb.study_apir.model;
 
 public class Product {
-    private int id;
+    private Long id;
     private String nome;
    
-    public Product(int id, String nome) {
+    public Product(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getNome() {
